asistente-backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
chromadb
PyMuPDF
python-dotenv
sqlalchemy
pydantic
